the image is also deleted from the vscode  after the product is deleted

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const { ValidationError } = require('sequelize');
 const Product = require('../models/product');
 const { body, validationResult } = require('express-validator');
+const fileHelper = require('../util/file')
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -134,6 +135,7 @@ exports.postEditProduct = (req, res, next) => {
       product.description = updatedDesc;
       if(image)
       {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save();
@@ -158,10 +160,20 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.findByIdAndRemove(prodId)
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
-    .catch(err => console.log(err));
+  Product.findById(prodId).then(product => {
+    if(!product)
+    {
+      return next(new Error('Product Not Found.'))
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.findByIdAndRemove(prodId)
+  })
+  .then(() => {
+    console.log('DESTROYED PRODUCT');
+    res.redirect('/admin/products');
+  })
+  .catch((err) => {
+    next(err);
+  })
+  
 };
